Refactor Action from interface to abstract class
