+++ src/main/java/carsharing/Main.java
@@ -8,6 +8,7 @@ import carsharing.database.dao.CustomerDao;
 import carsharing.util.CliParser;
 
 import java.util.Optional;
+import java.util.Scanner;
 
 public class Main {
     public static void main(String[] args) {
@@ -19,7 +20,7 @@ public class Main {
             CarDao carDao = new CarDao(dbManager);
             CustomerDao customerDao = new CustomerDao(dbManager);
 
-            MainActivity mainActivity = new MainActivity(companyDao, carDao, customerDao);
+            MainActivity mainActivity = new MainActivity(new Scanner(System.in), companyDao, carDao, customerDao);
             mainActivity.start();
 
             dbManager.closeConnection();
